First adapter, for testing
