+++ index.js
@@ -2,4 +2,4 @@ const server = require('./src/server.factory');
 const minimist = require('minimist');
 const args = minimist(process.argv.slice(2));
 
-server(args.port);
+server(args.port, args.adapter);
