+++ src/server.js
@@ -1,10 +1,15 @@
-module.exports = (deps) => (port) => {
+module.exports = (deps) => (port, adapter) => {
   // the adapter could be passed as a dependency too
   const { net } = deps;
   const connections = [];
+  const availableAdapters = {
+    rst1000: './adapter/rst1000'
+  }
+
   const server = net.createServer(conn => {
+    const adapter = require(availableAdapters[adapter])()
     // it will pass the adapter as device dependency
-    conn.device = require('./device')(conn);
+    conn.device = require('./device')({ adapter })(conn);
 
     connections.push(conn);
     console.log('new connection');
